# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.UserResponse'
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.UserResponse:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      nickname:
        type: string
      org_id:
        type: string
      org_name:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
...
  internal_interfaces_http_handler.BindPhoneRequest:
    description: 绑定手机号请求参数
    properties:
      code:
        description: 验证码（测试阶段可选）
        example: "123456"
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
    required:
    - phone
    type: object
...
  internal_interfaces_http_handler.LoginRequest:
    description: 用户登录请求参数
    properties:
      password:
        description: 密码
        example: admin123
        type: string
      username:
        description: 用户名/手机号
        example: "13800138000"
        type: string
    required:
    - password
    - username
    type: object
  internal_interfaces_http_handler.ResetPasswordRequest:
    description: 重置密码请求参数
    properties:
      code:
        description: 验证码
        example: "123456"
        type: string
      new_password:
        description: 新密码
        example: newpass123
        minLength: 8
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
    required:
    - code
    - new_password
    - phone
    type: object
  internal_interfaces_http_handler.SendCodeRequest:
    description: 发送短信验证码请求参数
    properties:
      phone:
        description: 手机号
        example: "13800138000"
        type: string
    required:
    - phone
    type: object
  internal_interfaces_http_handler.SendCodeResponse:
    description: 发送验证码成功响应
    properties:
      expire:
        description: 验证码有效期（秒）
        example: 300
        type: integer
      message:
        description: 响应消息
        example: 验证码已发送
        type: string
    type: object
  internal_interfaces_http_handler.WechatLoginRequest:
    description: 微信小程序登录请求参数
    properties:
      avatar:
        description: 用户头像（可选）
        example: https://example.com/avatar.jpg
        type: string
      code:
        description: 微信登录临时凭证
        example: wx_login_code_123
        type: string
      nickname:
        description: 用户昵称（可选）
        example: 张三
        type: string
    required:
    - code
    type: object
  internal_interfaces_http_handler.WechatLoginResponse:
    description: 微信登录响应，当need_bind_phone为true时需要绑定手机号
    properties:
      access_token:
        description: 访问令牌（临时）
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      expires_in:
        description: 过期时间（秒）
        example: 604800
        type: integer
      need_bind_phone:
        description: 是否需要绑定手机号
        example: true
        type: boolean
      refresh_token:
        description: 刷新令牌
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      token_type:
        description: 令牌类型
        example: Bearer
        type: string
      user:
        allOf:
        - $ref: '#/definitions/internal_interfaces_http_handler.WechatLoginUser'
        description: 用户信息
    type: object
  internal_interfaces_http_handler.WechatLoginUser:
    properties:
      id:
        description: 用户ID
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      nickname:
        description: 昵称
        example: 微信用户
        type: string
      phone:
        description: 手机号（可能为空）
        example: ""
        type: string
      role:
        description: 角色
        example: volunteer
        type: string
      status:
        description: 状态
        example: active
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@cntunyuan.org
    name: CNtunyuan Team
    url: https://github.com/Snowitty-Re/CNtunyuan
  description: 团圆寻亲志愿者系统后端 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Snowitty-Re/CNtunyuan
  title: 团圆寻亲志愿者系统 API
  version: 1.0.0
paths:
...
  /auth/admin-login:
    post:
      consumes:
      - application/json
      description: 管理员登录接口，与普通登录使用相同的认证逻辑，但通常用于管理后台
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.LoginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "403":
          description: 账号被禁用、锁定或无管理员权限
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 管理员登录
      tags:
      - Authentication
  /auth/bind-phone:
    post:
      consumes:
      - application/json
      description: 为微信登录用户绑定手机号，新用户首次登录时需要完成此步骤
      parameters:
      - default: Bearer <token>
        description: Bearer 访问令牌（微信登录获取的临时令牌）
        in: header
        name: Authorization
        required: true
        type: string
      - description: 绑定手机号请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.BindPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.LoginResponse'
              type: object
        "400":
          description: 参数错误或手机号格式不正确
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "409":
          description: 手机号已被其他账号绑定
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 绑定手机号
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录系统，返回访问令牌和用户信息
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.LoginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "403":
          description: 账号被禁用或锁定
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 用户登录
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 用户退出登录，使当前访问令牌失效
      parameters:
      - default: Bearer <token>
        description: Bearer 访问令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 用户登出
      tags:
      - Authentication
  /auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      parameters:
      - default: Bearer <token>
        description: Bearer 访问令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.UserResponse'
              type: object
        "401":
          description: 未授权或令牌无效
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 获取当前用户信息
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌，避免用户重新登录
      parameters:
      - description: 刷新令牌请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.LoginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 刷新令牌过期或无效
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "403":
          description: 账号被禁用
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 刷新访问令牌
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: 通过手机验证码重置密码，用于忘记密码场景
      parameters:
      - description: 重置密码请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 重置密码
      tags:
      - Authentication
  /auth/send-code:
    post:
      consumes:
      - application/json
      description: 向指定手机号发送短信验证码，用于手机号绑定、密码重置等场景
      parameters:
      - description: 发送验证码请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.SendCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.SendCodeResponse'
              type: object
        "400":
          description: 参数错误或手机号格式不正确
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "429":
          description: 发送过于频繁，请稍后再试
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误或短信服务异常
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 发送短信验证码
      tags:
      - Authentication
  /auth/wechat-login:
    post:
      consumes:
      - application/json
      description: 微信小程序登录接口，使用微信临时登录凭证换取系统访问令牌。新用户需要绑定手机号
      parameters:
      - description: 微信登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.WechatLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 需要绑定手机号（新用户）
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.WechatLoginResponse'
              type: object
        "400":
          description: 参数错误或微信登录凭证无效
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误或微信服务异常
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      summary: 微信小程序登录
      tags:
      - Authentication
...
securityDefinitions:
  Bearer:
    description: 请输入 JWT Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"